feat: add options (#297)
Added the option to specify a custom implementation of a HTTP client, where the default is based on Request (https://www.npmjs.com/package/request). This can be useful if discovery is required in an Electron application where one wish to benefit from the proxy configuration provided by using Electron's net.request.